requirements.txt
# ============================================
# 量化投资系统依赖文件
# ============================================

# 核心依赖 - 必需
requests>=2.28.0
pyyaml>=6.0
python-dateutil>=2.8.0
python-dotenv>=1.0.0

# 数据处理 - 核心功能需要 (兼容性版本)
pandas>=1.5.0,<2.0.0
numpy>=1.21.0,<2.0.0

# 数据源 - 可选但推荐
yfinance>=0.2.18
akshare>=1.17.0

# 高级数据源 - 可选 (需要token)
tushare>=1.2.89

# 任务调度 - 可选
schedule>=1.2.0

# 性能监控 - 可选
psutil>=5.9.0

# 开发和测试依赖
pytest>=7.0.0
pytest-cov>=4.0.0
black>=22.0.0
flake8>=5.0.0

# 技术分析 - 可选 (安装复杂，暂时注释)
# talib>=0.4.0

# 文档生成 - 可选
# sphinx>=5.0.0
# sphinx-rtd-theme>=1.0.0

# Jupyter支持 - 可选
# jupyter>=1.0.0
# matplotlib>=3.5.0
# seaborn>=0.11.0
# plotly>=5.0.0
